backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langgraph
langchain
langchain-openai
llama-index
llama-index-vector-stores-postgres
psycopg2-binary
httpx
pydantic
anthropic
langchain-google-genai
ollama
pymupdf>=1.24.0
pymupdf4llm>=0.0.10
tiktoken>=0.7.0
click>=8.0.0
fpdf2>=2.8.0
pyyaml>=6.0
